requirements.txt
Django==3.2.8
psycopg2==2.8.4
psycopg2-binary==2.9.3
python-dotenv==0.19.2
pandas==1.4.0
tqdm==4.62.3
python-slugify==5.0.2
requests==2.27.1
names==0.3.0
Pillow==9.0.1

# tests
pytest==6.2.5
pytest-sugar==0.9.4
pytest-django==4.1.0

# lint
black==20.8b1

numpy~=1.22.1